Introducing Core.Model.Rotations namespace.
