+++ RubbikCube.Core/Model/Cube.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 using System.Text;
+using RubiksCube.Core.Model.Rotations;
 
 namespace RubiksCube.Core.Model
 {
