+++ RubbikCube.Core/Model/Face.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
+using RubiksCube.Core.Model.Rotations;
 
 namespace RubiksCube.Core.Model
 {
