+++ RubbikCube.Core/Model/FaciePositionType.cs
@@ -1,4 +1,6 @@
-﻿namespace RubiksCube.Core.Model
+﻿using RubiksCube.Core.Model.Rotations;
+
+namespace RubiksCube.Core.Model
 {
     public enum FaciePositionType
     {
