+++ RubbikCube.Core/Model/Rotations/DownWholeFaceRotation.cs
@@ -1,27 +1,27 @@
 ﻿using System.Collections.Generic;
 
-namespace RubiksCube.Core.Model
+namespace RubiksCube.Core.Model.Rotations
 {
-    public class LeftWholeFaceRotation : FaceRotation
+    public class DownWholeFaceRotation : FaceRotation
     {
-        public LeftWholeFaceRotation(double angle, uint times)
+        public DownWholeFaceRotation(double angle, uint times)
             : base(Clockwise, angle, times)
         {
         }
 
         public override double[,] GetRotationMatrix(double angle)
         {
-            return rotationMatrixFactory.CreateXRotationMatrix(angle);
+            return RotationMatrixFactory.CreateYRotationMatrix(angle);
         }
 
         protected override IEnumerable<FaceType> GetImpactedFaceTypes()
         {
-            return new[] { FaceType.Front, FaceType.Left, FaceType.Back, FaceType.Right };
+            return new[] { FaceType.Front, FaceType.Top, FaceType.Back, FaceType.Bottom };
         }
 
         protected override IEnumerable<Facie> GetImpactedFacies(Face face)
         {
-            return face.GetRowFacies(RotationType.All);
+            return face.GetColumnFacies(RotationType.All);
         }
 
         protected override void Move(Cube cube, FaceType faceType, Facie facie, bool isPositiveRotation)
@@ -32,48 +32,48 @@ namespace RubiksCube.Core.Model
             switch (faceType)
             {
                 case FaceType.Front:
-                    cube.LeftFace.Facies.Add(facie);
+                    FlipPosition(facie);
+                    cube.BottomFace.Facies.Add(facie);
                     break;
-                case FaceType.Right:
+                case FaceType.Top:
                     cube.FrontFace.Facies.Add(facie);
                     break;
                 case FaceType.Back:
                     FlipPosition(facie);
-                    cube.RightFace.Facies.Add(facie);
+                    cube.TopFace.Facies.Add(facie);
                     break;
-                case FaceType.Left:
-                    FlipPosition(facie);
+                case FaceType.Bottom:
                     cube.BackFace.Facies.Add(facie);
                     break;
             }
         }
 
-        // TODO: same method for top/bottom face rotation -> create abstraction
+        // TODO: same method for rignt/left face rotation -> create abstraction
         private static void FlipPosition(Facie facie)
         {
             if (facie.FaciePosition == FaciePositionType.LeftTop)
             {
-                facie.FaciePosition = FaciePositionType.RightTop;
+                facie.FaciePosition = FaciePositionType.LeftBottom;
             }
-            else if (facie.FaciePosition == FaciePositionType.RightTop)
+            else if (facie.FaciePosition == FaciePositionType.LeftBottom)
             {
                 facie.FaciePosition = FaciePositionType.LeftTop;
             }
-            else if (facie.FaciePosition == FaciePositionType.LeftMiddle)
+            else if (facie.FaciePosition == FaciePositionType.MiddleTop)
             {
-                facie.FaciePosition = FaciePositionType.RightMiddle;
+                facie.FaciePosition = FaciePositionType.MiddleBottom;
             }
-            else if (facie.FaciePosition == FaciePositionType.RightMiddle)
+            else if (facie.FaciePosition == FaciePositionType.MiddleBottom)
             {
-                facie.FaciePosition = FaciePositionType.LeftMiddle;
+                facie.FaciePosition = FaciePositionType.MiddleTop;
             }
-            else if (facie.FaciePosition == FaciePositionType.LeftBottom)
+            else if (facie.FaciePosition == FaciePositionType.RightTop)
             {
                 facie.FaciePosition = FaciePositionType.RightBottom;
             }
             else if (facie.FaciePosition == FaciePositionType.RightBottom)
             {
-                facie.FaciePosition = FaciePositionType.LeftBottom;
+                facie.FaciePosition = FaciePositionType.RightTop;
             }
         }
     }
