+++ RubbikCube.Core/Model/Rotations/RightFaceRotation.cs
@@ -1,17 +1,17 @@
 ﻿using System.Collections.Generic;
 
-namespace RubiksCube.Core.Model
+namespace RubiksCube.Core.Model.Rotations
 {
-    public class UpWholeFaceRotation : FaceRotation
+    public class RightFaceRotation : FaceRotation
     {
-        public UpWholeFaceRotation(double angle, uint times)
-            : base(CounterClockwise, angle, times)
+        public RightFaceRotation(string way, double angle, uint times) 
+            : base(way, angle, times)
         {
         }
 
         public override double[,] GetRotationMatrix(double angle)
         {
-            return rotationMatrixFactory.CreateYRotationMatrix(angle);
+            return RotationMatrixFactory.CreateYRotationMatrix(angle);
         }
 
         protected override IEnumerable<FaceType> GetImpactedFaceTypes()
@@ -21,7 +21,7 @@ namespace RubiksCube.Core.Model
 
         protected override IEnumerable<Facie> GetImpactedFacies(Face face)
         {
-            return face.GetColumnFacies(RotationType.All);
+            return face.GetColumnFacies(RotationType.Third);
         }
 
         protected override void Move(Cube cube, FaceType faceType, Facie facie, bool isPositiveRotation)
@@ -29,21 +29,51 @@ namespace RubiksCube.Core.Model
             var sourceFace = cube.Find(faceType);
             sourceFace.Facies.Remove(facie);
 
-            switch (faceType)
+            switch(faceType)
             {
                 case FaceType.Front:
-                    cube.TopFace.Facies.Add(facie);
+                    if (isPositiveRotation)
+                    {
+                        cube.TopFace.Facies.Add(facie);
+                    }
+                    else
+                    {
+                        FlipPosition(facie);
+                        cube.BottomFace.Facies.Add(facie);
+                    }
                     break;
                 case FaceType.Top:
-                    FlipPosition(facie);
-                    cube.BackFace.Facies.Add(facie);
+                    if (isPositiveRotation)
+                    {
+                        FlipPosition(facie);
+                        cube.BackFace.Facies.Add(facie);
+                    }
+                    else
+                    {
+                        cube.FrontFace.Facies.Add(facie);
+                    }
                     break;
                 case FaceType.Back:
-                    cube.BottomFace.Facies.Add(facie);
+                    if (isPositiveRotation)
+                    {
+                        cube.BottomFace.Facies.Add(facie);
+                    }
+                    else
+                    {
+                        FlipPosition(facie);
+                        cube.TopFace.Facies.Add(facie);
+                    }
                     break;
                 case FaceType.Bottom:
-                    FlipPosition(facie);
-                    cube.FrontFace.Facies.Add(facie);
+                    if (isPositiveRotation)
+                    {
+                        FlipPosition(facie);
+                        cube.FrontFace.Facies.Add(facie);
+                    }
+                    else
+                    {
+                        cube.BackFace.Facies.Add(facie);
+                    }
                     break;
             }
         }
