+++ RubbikCube.Core/Model/Rotations/RotationFactory.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace RubiksCube.Core.Model
+namespace RubiksCube.Core.Model.Rotations
 {
     public interface IRotationFactory
     {
