+++ RubbikCube.Core/RotationAttribute.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using RubiksCube.Core.Model;
+using RubiksCube.Core.Model.Rotations;
 
 namespace RubiksCube.Core
 {
