+++ RubbikCube.Core/RubiksCube.Core.csproj
@@ -41,27 +41,27 @@
   <ItemGroup>
     <Compile Include="EnumerableExtensions.cs" />
     <Compile Include="Model\Color.cs" />
-    <Compile Include="Model\ForwardFaceRotation.cs" />
-    <Compile Include="Model\LeftWholeFaceRotation.cs" />
-    <Compile Include="Model\RightWholeFaceRotation.cs" />
-    <Compile Include="Model\RotationFactory.cs" />
-    <Compile Include="Model\TopFaceRotation.cs" />
-    <Compile Include="Model\LeftFaceRotation.cs" />
+    <Compile Include="Model\Rotations\ForwardFaceRotation.cs" />
+    <Compile Include="Model\Rotations\LeftWholeFaceRotation.cs" />
+    <Compile Include="Model\Rotations\RightWholeFaceRotation.cs" />
+    <Compile Include="Model\Rotations\RotationFactory.cs" />
+    <Compile Include="Model\Rotations\TopFaceRotation.cs" />
+    <Compile Include="Model\Rotations\LeftFaceRotation.cs" />
     <Compile Include="Model\Facie.cs" />
-    <Compile Include="Model\BottomFaceRotation.cs" />
-    <Compile Include="Model\FaceRotation.cs" />
+    <Compile Include="Model\Rotations\BottomFaceRotation.cs" />
+    <Compile Include="Model\Rotations\FaceRotation.cs" />
     <Compile Include="Model\Cube.cs" />
     <Compile Include="Model\Face.cs" />
     <Compile Include="Model\FaceType.cs" />
     <Compile Include="Model\FaciePositionType.cs" />
-    <Compile Include="Model\RightFaceRotation.cs" />
-    <Compile Include="Model\RotationType.cs" />
+    <Compile Include="Model\Rotations\RightFaceRotation.cs" />
+    <Compile Include="Model\Rotations\RotationType.cs" />
     <Compile Include="Model\FaceFactory.cs" />
-    <Compile Include="Model\RotationMatrixFactory.cs" />
+    <Compile Include="Model\Rotations\RotationMatrixFactory.cs" />
     <Compile Include="Guard.cs" />
     <Compile Include="MatrixHelper.cs" />
-    <Compile Include="Model\DownWholeFaceRotation.cs" />
-    <Compile Include="Model\UpWholeFaceRotation.cs" />
+    <Compile Include="Model\Rotations\DownWholeFaceRotation.cs" />
+    <Compile Include="Model\Rotations\UpWholeFaceRotation.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="RotationAttribute.cs" />
     <Compile Include="ReflectionHelper.cs" />
